Refactor vortex tokenizer module for better cohesion
Extracted `SpecialTokens`, `ChatRole`, and `ChatMessage` to `types.rs`.
Extracted chat template logic to `template.rs`.
Reduced `vortex/src/tokenizer/mod.rs` from ~747 lines to ~200 lines.
Enforced stricter visibility for internal modules.
